# perf(expr): apply auto-vectorization and remove explicit SIMDs