Following #554, this PR applies its auto-vectorization evaluation to almost all operations. 
Since it has outperformed the explicit SIMD evaluation, we decide to remove the latter.

This PR also adds more benchmarks for many kinds of types and expressions.
Here are some results on my M1 Pro macOS: (chunk size = 4096)
| bench                | old       | new       | change   | Status |
| -------------------- | --------- | --------- | -------- | ------ |
| add(i32,i32)         | 26.282 µs | 1.2417 µs | -95.270% | ✅      |
| mul(i32,i32)         | 26.389 µs | 1.2534 µs | -95.230% | ✅      |
| eq(i32,i32)          | 26.325 µs | 1.1615 µs | -95.546% | ✅      |
| gt(i32,i32)          | 26.204 µs | 1.1588 µs | -95.595% | ✅      |
| add(f64,f64)         | 26.204 µs | 1.5329 µs | -94.128% | ✅      |
| mul(f64,f64)         | 26.372 µs | 1.5132 µs | -94.211% | ✅      |
| eq(f64,f64)          | 26.307 µs | 2.8696 µs | -89.094% | ✅      |
| gt(f64,f64)          | 27.407 µs | 3.8167 µs | -86.041% | ✅      |
| add(decimal,decimal) | 35.499 µs | 14.474 µs | -59.255% | ❗️      |
| mul(decimal,decimal) | 35.248 µs | 12.559 µs | -64.195% | ❗️      |
| eq(decimal,decimal)  | 35.866 µs | 13.762 µs | -61.290% | ❗️      |
| gt(decimal,decimal)  | 35.844 µs | 13.831 µs | -60.907% | ❗️      |
| and(bool,bool)       | 26.306 µs | 21.563 µs | -17.726% | ❗️      |
| not(bool)            | 18.610 µs | 160.53 ns | -99.132% | ✅      |
| sum(i32)             | 9.7897 µs | 299.58 ns | -96.955% | ✅      |
| max(i32)             | 9.7429 µs | 9.7349 µs | --       | ❗️      |
| first(i32)           | 5.6818 ns | 5.7933 ns | --       | ✅      |
| count(i32)           | 14.306 ns | 13.695 ns | -4.7650% | ✅      |
| sum(f64)             | 9.6450 µs | 3.7473 µs | -61.034% | ✅      |
| max(f64)             | 17.905 µs | 17.129 µs | --       | ❗️      |
| first(f64)           | 5.9505 ns | 6.0258 ns | --       | ✅      |
| count(f64)           | 14.620 ns | 13.528 ns | -6.8730% | ✅      |
| sum(decimal)         | 26.092 µs | 29.657 µs | +13.737% | ❗️      |
| max(decimal)         | 23.591 µs | 23.709 µs | --       | ❗️      |
| first(decimal)       | 6.4617 ns | 6.5106 ns | --       | ✅      |
| count(decimal)       | 14.313 ns | 13.573 ns | -5.3045% | ✅      |
| cast(i32->f64)       | 19.301 µs | 558.22 ns | -97.118% | ✅      |
| cast(f64->decimal)   | 123.01 µs | 124.60 µs | --       | ❗️      |
| cast(i32->string)    | 85.323 µs | 86.483 µs | --       | ❗️      |

✅: almost optimal
❗️: needs to improve

TPC-H results:

| time(s) | old   | new   |
| ------- | ----- | ----- |
| Q1      | 1.871 | 1.580 |
| Q3      | 0.419 | 0.434 |
| Q5      | 0.930 | 0.853 |
| Q6      | 0.729 | 0.469 |
| Q10     | 0.684 | 0.672 |